requirements.txt
camelot-py==0.9.0
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
deprecation==2.1.0
et-xmlfile==1.1.0
fonttools==4.39.4
img2pdf==0.4.4
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
numpy==1.24.3
opencv-python==4.7.0.72
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
pdf2image==1.16.3
pdfminer.six==20221105
pika==1.3.2
pikepdf==7.2.0
Pillow==9.5.0
pycparser==2.21
pydantic==1.10.7
pyparsing==3.0.9
PyPDF2==2.12.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
typing_extensions==4.5.0
tzdata==2023.3
